+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Button System
 - RequestAssets Model
 - Permission System
+- Notification Send System
 
 ### Changed